+++ config.js
@@ -1,0 +1,5 @@
+var config = {};
+
+config.sdkGenURL = 'https://mobilesdkgen.ng.bluemix.net/sdkgen/api/generator/';
+
+module.exports = config;